+++ src/styles/character.css
@@ -1,3 +1,10 @@
+.character-name {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-items: center;
+  padding-bottom: 30px;
+}
 
 .character-name .name {
   font-size: 28px;
